+++ entities/position_test.go
@@ -27,8 +27,8 @@ func initPool() (*Pool, int, int) {
 		panic(err)
 	}
 
-	tickSpacing := constants.TickSpacings[constants.FeeLow]
-	p, err := NewPool(DAI, USDC, constants.FeeLow, poolSqrtRatioStart, big.NewInt(0), poolTickCurrent, nil)
+	tickSpacing := constants.TickSpacings[uint64(constants.FeeLow)]
+	p, err := NewPool(DAI, USDC, uint64(constants.FeeLow), poolSqrtRatioStart, big.NewInt(0), poolTickCurrent, nil)
 	if err != nil {
 		panic(err)
 	}
@@ -197,7 +197,7 @@ func TestMintAmountsWithSlippage(t *testing.T) {
 	slippageTolerance = entities.NewPercent(big.NewInt(5), big.NewInt(100))
 
 	// is correct for pool at min price
-	minPricePool, err := NewPool(DAI, USDC, constants.FeeLow, utils.MinSqrtRatio, big.NewInt(0), utils.MinTick, nil)
+	minPricePool, err := NewPool(DAI, USDC, uint64(constants.FeeLow), utils.MinSqrtRatio, big.NewInt(0), utils.MinTick, nil)
 	assert.NoError(t, err)
 	p, err = NewPosition(minPricePool, B100e18,
 		NearestUsableTick(poolTickCurrent, tickSpacing)+tickSpacing,
@@ -209,7 +209,7 @@ func TestMintAmountsWithSlippage(t *testing.T) {
 	assert.Equal(t, "0", amount1.String())
 
 	// is correct for pool at max price
-	maxPricePool, err := NewPool(DAI, USDC, constants.FeeLow, new(big.Int).Sub(utils.MaxSqrtRatio, big.NewInt(1)), big.NewInt(0), utils.MaxTick-1, nil)
+	maxPricePool, err := NewPool(DAI, USDC, uint64(constants.FeeLow), new(big.Int).Sub(utils.MaxSqrtRatio, big.NewInt(1)), big.NewInt(0), utils.MaxTick-1, nil)
 	assert.NoError(t, err)
 	p, err = NewPosition(maxPricePool, B100e18,
 		NearestUsableTick(poolTickCurrent, tickSpacing)+tickSpacing,
@@ -296,7 +296,7 @@ func TestBurnAmountsWithSlippage(t *testing.T) {
 	slippageTolerance = entities.NewPercent(big.NewInt(5), big.NewInt(100))
 
 	// is correct for pool at min price
-	minPricePool, err := NewPool(DAI, USDC, constants.FeeLow, utils.MinSqrtRatio, big.NewInt(0), utils.MinTick, nil)
+	minPricePool, err := NewPool(DAI, USDC, uint64(constants.FeeLow), utils.MinSqrtRatio, big.NewInt(0), utils.MinTick, nil)
 	assert.NoError(t, err)
 	p, err = NewPosition(minPricePool, B100e18,
 		NearestUsableTick(poolTickCurrent, tickSpacing)+tickSpacing,
@@ -309,7 +309,7 @@ func TestBurnAmountsWithSlippage(t *testing.T) {
 	assert.Equal(t, "0", amount1.String())
 
 	// is correct for pool at max price
-	maxPricePool, err := NewPool(DAI, USDC, constants.FeeLow, new(big.Int).Sub(utils.MaxSqrtRatio, big.NewInt(1)), big.NewInt(0), utils.MaxTick-1, nil)
+	maxPricePool, err := NewPool(DAI, USDC, uint64(constants.FeeLow), new(big.Int).Sub(utils.MaxSqrtRatio, big.NewInt(1)), big.NewInt(0), utils.MaxTick-1, nil)
 	assert.NoError(t, err)
 	p, err = NewPosition(maxPricePool, B100e18,
 		NearestUsableTick(poolTickCurrent, tickSpacing)+tickSpacing,
