+++ utils/compute_pool_address_test.go
@@ -13,7 +13,7 @@ func TestComputePoolAddress(t *testing.T) {
 	factoryAddress := common.HexToAddress("0x1111111111111111111111111111111111111111")
 	tokenA := entities.NewToken(1, common.HexToAddress("0xA0b86991c6218b36c1d19D4a2e9Eb0cE3606eB48"), 18, "USDC", "USD Coin")
 	tokenB := entities.NewToken(1, common.HexToAddress("0x6B175474E89094C44Da98b954EedeAC495271d0F"), 18, "DAI", "Dai Stablecoin")
-	result, err := ComputePoolAddress(factoryAddress, tokenA, tokenB, constants.FeeLow, "")
+	result, err := ComputePoolAddress(factoryAddress, tokenA, tokenB, uint64(constants.FeeLow), "")
 	if err != nil {
 		panic(err)
 	}
@@ -21,11 +21,11 @@ func TestComputePoolAddress(t *testing.T) {
 
 	USDC := entities.NewToken(1, common.HexToAddress("0xA0b86991c6218b36c1d19D4a2e9Eb0cE3606eB48"), 18, "USDC", "USD Coin")
 	DAI := entities.NewToken(1, common.HexToAddress("0x6B175474E89094C44Da98b954EedeAC495271d0F"), 18, "DAI", "Dai Stablecoin")
-	resultA, err := ComputePoolAddress(factoryAddress, USDC, DAI, constants.FeeLow, "")
+	resultA, err := ComputePoolAddress(factoryAddress, USDC, DAI, uint64(constants.FeeLow), "")
 	if err != nil {
 		panic(err)
 	}
-	resultB, err := ComputePoolAddress(factoryAddress, DAI, USDC, constants.FeeLow, "")
+	resultB, err := ComputePoolAddress(factoryAddress, DAI, USDC, uint64(constants.FeeLow), "")
 	if err != nil {
 		panic(err)
 	}
